Partial Commit: Moved CheckBox related contents to demo package
