+++ livedirsfx-demo/build.gradle
@@ -2,6 +2,7 @@ group 'org.fxmisc.livedirs'
 
 dependencies {
     compile project(":livedirsfx")
+    compile group: 'org.reactfx', name: 'reactfx', version: '2.0-M4u1'
 }
 
 task fatJar(type: Jar, dependsOn: classes) {
