+++ livedirsfx-demo/src/main/java/org/fxmisc/livedirs/demo/checkbox/Recursive.java
@@ -1,4 +1,4 @@
-package org.fxmisc.livedirs;
+package org.fxmisc.livedirs.demo.checkbox;
 
 /**
  * Allows recursive {@link FunctionalInterface} calls.
