+++ livedirsfx-demo/src/main/java/org/fxmisc/livedirs/demo/checkbox/TreeCellFactories.java
@@ -1,4 +1,4 @@
-package org.fxmisc.livedirs;
+package org.fxmisc.livedirs.demo.checkbox;
 
 import javafx.scene.control.TreeCell;
 import javafx.scene.control.TreeView;
@@ -7,7 +7,7 @@ import javafx.util.Callback;
 import java.util.function.Function;
 
 /**
- * A class that provides cell factories to help display a {@link PathItem}
+ * A class that provides cell factories to help display a {@link org.fxmisc.livedirs.PathItem}
  * as though it extended a subclass of {@link javafx.scene.control.TreeItem}
  * or to provide additional functionality not present in the normal cell factory.
  */
@@ -18,9 +18,9 @@ public class TreeCellFactories {
     }
 
     public static <C extends CheckBoxContent> Callback<TreeView<C>, TreeCell<C>> checkBoxFactory(
-            Function<CheckBoxContent, String> stringConverter
+            Function<C, String> stringConverter
     ) {
-        return (view) -> new CheckBoxTreeCell(stringConverter);
+        return (view) -> new CheckBoxTreeCell<>(stringConverter);
     }
 
 }
